changed type from long to double for elemw plus and mult

+++ src/main/java/org/apache/sysml/hops/estim/EstimatorMatrixHistogram.java
@@ -101,21 +101,21 @@ public class EstimatorMatrixHistogram extends SparsityEstimator
 			case MM:
 				return estimInternMM(h1, h2);
 			case MULT: {
-				final long N1 = h1.getNonZeros();
-				final long N2 = h2.getNonZeros();
+				final double N1 = h1.getNonZeros();
+				final double N2 = h2.getNonZeros();
 				final long scale = IntStream.range(0, h1.getCols())
 					.mapToLong(j -> (long)h1.cNnz[j] * h2.cNnz[j]).sum();
 				return IntStream.range(0, h1.getRows())
-					.mapToLong(i -> (long)h1.rNnz[i] * h2.rNnz[i] * scale / N1 / N2) //collisions
+					.mapToDouble(i -> (long)h1.rNnz[i] * h2.rNnz[i] * scale / N1 / N2) //collisions
 					.sum() / msize;
 			}
 			case PLUS: {
-				final long N1 = h1.getNonZeros();
-				final long N2 = h2.getNonZeros();
+				final double N1 = h1.getNonZeros();
+				final double N2 = h2.getNonZeros();
 				final long scale = IntStream.range(0, h1.getCols())
 					.mapToLong(j -> (long)h1.cNnz[j] * h2.cNnz[j]).sum();
 				return IntStream.range(0, h1.getRows())
-					.mapToLong(i -> (long)h1.rNnz[i] + h2.rNnz[i] //all minus collisions
+					.mapToDouble(i -> (long)h1.rNnz[i] + h2.rNnz[i] //all minus collisions
 						- (long)h1.rNnz[i] * h2.rNnz[i] * scale / N1 / N2)
 					.sum() / msize;
 			}
@@ -354,12 +354,12 @@ public class EstimatorMatrixHistogram extends SparsityEstimator
 		}
 		
 		private static MatrixHistogram deriveMultHistogram(MatrixHistogram h1, MatrixHistogram h2) {
-			final long N1 = h1.getNonZeros();
-			final long N2 = h2.getNonZeros();
-			final long scaler = IntStream.range(0, h1.getCols())
-				.mapToLong(j -> (long)h1.cNnz[j] * h2.cNnz[j]).sum();
-			final long scalec = IntStream.range(0, h1.getRows())
-				.mapToLong(j -> (long)h1.rNnz[j] * h2.rNnz[j]).sum();
+			final double N1 = h1.getNonZeros();
+			final double N2 = h2.getNonZeros();
+			final double scaler = IntStream.range(0, h1.getCols())
+				.mapToDouble(j -> (long)h1.cNnz[j] * h2.cNnz[j]).sum();
+			final double scalec = IntStream.range(0, h1.getRows())
+				.mapToDouble(j -> (long)h1.rNnz[j] * h2.rNnz[j]).sum();
 			int rMaxNnz = 0, cMaxNnz = 0;
 			Random rn = new Random();
 			int[] rNnz = new int[h1.getRows()];
